Authorization implemented

+++ src/api/notes/handler.js
@@ -1,10 +1,5 @@
-/* eslint-disable max-len */
-/* eslint-disable consistent-return */
-/* eslint-disable no-undef */
-
 const ClientError = require('../../exceptions/ClientError');
 
-/* eslint-disable no-underscore-dangle */
 class NotesHandler {
   constructor(service, validator) {
     this._service = service;
@@ -21,8 +16,11 @@ class NotesHandler {
     try {
       this._validator.validateNotePayLoad(request.payload);
       const { title = 'untitled', body, tags } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
 
-      const noteId = await this._service.addNote({ title, body, tags });
+      const noteId = await this._service.addNote({
+        title, body, tags, owner: credentialId,
+      });
 
       const response = h.response({
         status: 'success',
@@ -54,8 +52,9 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -67,6 +66,9 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
       return {
         status: 'success',
@@ -99,7 +101,9 @@ class NotesHandler {
     try {
       this._validator.validateNotePayLoad(request.payload);
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
 
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.editNoteById(id, request.payload);
 
       return {
@@ -130,6 +134,9 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: owner } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, owner);
       await this._service.deleteNoteById(id);
 
       return {
